Added two new tools
averageDist.py Can do a number of calculations, but was made to
calculate the average divergence within a subtree. Rerooting is an
experimental feature. Currently only supports rooted trees.

lenSwap.py Takes a current Newick tree, a .fit file from HyPhy,
and a parameter from the model and swaps the branch lengths in the newick
with the value of that parameter. Can match one of multiple parameters, used
mostly with t and syn. These are passed as variable number arguments at the
end of the command.